\name{bife-package}
\alias{bife-package}
\docType{package}
\title{
\packageTitle{bife}
}
\description{
\packageDescription{bife}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{bife}
\packageIndices{bife}
}
\references{
Hahn, J., and W. Newey (2004). "Jackknife and analytical bias reduction 
for nonlinear panel models." Econometrica 72(4), 1295-1319.

Stammann, A., F. Heiss, and D. McFadden (2016). "Estimating Fixed Effects 
Logit Models with Large Panel Data". Working paper.
}
\keyword{ package }

