% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.bife.R
\name{coef.bife}
\alias{coef.bife}
\title{Extract Model Coefficients}
\usage{
\method{coef}{bife}(object, corrected = TRUE, fixed = FALSE, ...)
}
\arguments{
\item{object}{an object of class \code{bife}.}

\item{corrected}{an optional logical flag that specifies whether bias-corrected or uncorrected
coefficients are displayed. Default is \code{TRUE} (bias-corrected).}

\item{fixed}{an optional logical flag that specifies whether the structural parameters or the fixed 
effects are displayed. Default is \code{FALSE} (structural parameters).}

\item{...}{other arguments}
}
\value{
The function \code{coef.bife} returns a named vector of coefficients.
}
\description{
\code{coef.bife} is a generic function which extracts
model coefficients from objects returned by \code{bife}.
}
\author{
Amrei Stammann, Daniel Czarnowske, Florian Heiss, Daniel McFadden
}
\seealso{
\code{\link{bife}}
}

