% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{coef.bifeAPEs}
\alias{coef.bifeAPEs}
\title{Extract estimates of average partial effects}
\usage{
\method{coef}{bifeAPEs}(object, ...)
}
\arguments{
\item{object}{an object of class \code{"APEs"}.}

\item{...}{other arguments.}
}
\value{
The function \code{\link{coef.bifeAPEs}} returns a named vector of estimates of the average 
partial effects.
}
\description{
\code{\link{coef.bifeAPEs}} is a generic function which extracts estimates of the average partial 
effects from objects returned by \code{\link{get_APEs}}.
}
\seealso{
\code{\link{get_APEs}}
}
