% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{fitted.bife}
\alias{fitted.bife}
\title{Extract \code{bife} fitted values}
\usage{
\method{fitted}{bife}(object, ...)
}
\arguments{
\item{object}{an object of class \code{"bife"}.}

\item{...}{other arguments.}
}
\value{
The function \code{\link{fitted.bife}} returns a vector of fitted values.
}
\description{
\code{\link{fitted.bife}} is a generic function which extracts fitted values from an object 
returned by \code{\link{bife}}.
}
\seealso{
\code{\link{bife}}
}
