% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{logLik.bife}
\alias{logLik.bife}
\title{Extract log-likelihood}
\usage{
\method{logLik}{bife}(object, ...)
}
\arguments{
\item{object}{an object of class \code{"bife"}.}

\item{...}{other arguments.}
}
\value{
The function \code{\link{logLik.bife}} returns the sum of the log-likelihood.
}
\description{
\code{\link{logLik.bife}} extracts the sum of the log-likelihood from an object 
returned by \code{\link{bife}}.
}
\seealso{
\code{\link{bife}}
}
