% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_bife.R
\name{print.bife}
\alias{print.bife}
\title{Print \code{bife}}
\usage{
\method{print}{bife}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{an object of class \code{bife}.}

\item{digits}{integer indicating the number of decimal places. Default is 
\code{max(3, getOption("digits") - 3)}.}

\item{...}{other arguments}
}
\description{
\code{print.bife} is a generic function which displays some minimal
information from objects returned by \code{bife}.
}
\seealso{
\code{\link{bife}}
}
\author{
Amrei Stammann, Daniel Czarnowske, Florian Heiss, Daniel McFadden
}
