% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{print.bifeAPEs}
\alias{print.bifeAPEs}
\title{Print \code{bifeAPEs}}
\usage{
\method{print}{bifeAPEs}(x, digits = max(3L, getOption("digits") - 3L),
  ...)
}
\arguments{
\item{x}{an object of class \code{"bifeAPEs"}.}

\item{digits}{unsigned integer indicating the number of decimal places. Default is 
\code{max(3L, getOption("digits") - 3L)}.}

\item{...}{other arguments.}
}
\description{
\code{\link{print.bifeAPEs}} is a generic function which displays some minimal information from 
objects returned by \code{\link{get_APEs}}.
}
\seealso{
\code{\link{get_APEs}}
}
