% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.bife.R
\name{print.summary.bife}
\alias{print.summary.bife}
\title{Print \code{summary.bife}}
\usage{
\method{print}{summary.bife}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{an object of class \code{summary.bife}.}

\item{digits}{integer indicating the number of decimal places. Default is 
\code{max(3, getOption("digits") - 3)}.}

\item{...}{other arguments}
}
\description{
\code{print.summary.bife} is a generic function which displays summary statistics
from objects returned by \code{summary.bife}.
}
\author{
Amrei Stammann, Daniel Czarnowske, Florian Heiss, Daniel McFadden
}
\seealso{
\code{\link{bife}}
}

