% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{print.summary.bifeAPEs}
\alias{print.summary.bifeAPEs}
\title{Print \code{summary.bifeAPEs}}
\usage{
\method{print}{summary.bifeAPEs}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{an object of class \code{"summary.bifeAPEs"}.}

\item{digits}{unsigned integer indicating the number of decimal places. Default is 
\code{max(3L, getOption("digits") - 3L)}.}

\item{...}{other arguments.}
}
\description{
\code{\link{print.summary.bifeAPEs}} is a generic function which displays summary statistics from 
objects returned by \code{\link{summary.bifeAPEs}}.
}
\seealso{
\code{\link{get_APEs}}
}
