% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psid.R
\docType{data}
\name{psid}
\alias{psid}
\title{Female labor force participation - "Panel Study of Income Dynamics"}
\format{A data frame with 13,149 rows:
\describe{
  \item{ID}{individual identifier}
  \item{LFP}{labor force participation}
  \item{KID1}{# of kids 0-2}
  \item{KID2}{# of kids 3-5}
  \item{KID3}{# of kids 6-17}
  \item{INCH}{income husband}
  \item{AGE}{age of woman}
  \item{TIME}{time identifier}}}
\usage{
psid
}
\description{
The sample was obtained from the "Panel Study of Income Dynamics" and
contains information about \eqn{N = 1461} women that were observed over
\eqn{T = 9} years.
}
\references{
Hyslop, D. (1999). "State Dependence, Serial Correlation and Heterogeneity in Intertemporal Labor Force Participation of Married Women". Econometrica 67(6), 1255-1294.
}
\seealso{
\code{\link{bife}}
}
\keyword{datasets}
