% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.acs.R
\docType{data}
\name{results.acs}
\alias{results.acs}
\title{Results "American Community Survey"}
\format{A named matrix with 4 rows and 4 columns.}
\usage{
results.acs
}
\description{
Results reported in the vignette.
}
\seealso{
\code{\link{bife}}
}
\keyword{datasets}

