% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_acs.R
\docType{data}
\name{results_acs}
\alias{results_acs}
\title{Results "American Community Survey"}
\format{A named matrix with 4 rows and 4 columns.}
\usage{
results_acs
}
\description{
Results reported in the vignette.
}
\seealso{
\code{\link{bife}}
}
\keyword{datasets}
