% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_psid.R
\docType{data}
\name{results_psid}
\alias{results_psid}
\title{Results "Panel Study of Income Dynamics"}
\format{A named matrix with 6 rows and 4 columns.}
\usage{
results_psid
}
\description{
Results reported in the vignette.
}
\seealso{
\code{\link{bife}}
}
\keyword{datasets}
