% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_bife.R
\name{summary.bife}
\alias{summary.bife}
\title{Summarizing Binary Choice Models with Fixed Effects}
\usage{
\method{summary}{bife}(object, corrected = TRUE, fixed = FALSE, ...)
}
\arguments{
\item{object}{an object of class \code{bife}.}

\item{corrected}{an optional logical flag that specifies 
whether bias-corrected or uncorrected coefficients 
are displayed. Default is \code{TRUE} (bias-corrected).}

\item{fixed}{an optional logical flag that specifies whether 
only structural parameters or all coefficients
(structural parameters and fixed effects) are displayed.
Default is \code{FALSE} (only structural parameters).}

\item{...}{other arguments}
}
\value{
Returns an object of class \code{summary.bife} which is a list of summary
statistics of \code{object}.
}
\description{
Summary statistics for objects of class \code{bife}.
}
\seealso{
\code{\link{bife}}
}
\author{
Amrei Stammann, Daniel Czarnowske, Florian Heiss, Daniel McFadden
}
