% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{summary.bifeAPEs}
\alias{summary.bifeAPEs}
\title{Summarizing models of class \code{bifeAPEs}}
\usage{
\method{summary}{bifeAPEs}(object, ...)
}
\arguments{
\item{object}{an object of class \code{"bifeAPEs"}.}

\item{...}{other arguments.}
}
\value{
Returns an object of class \code{"summary.bifeAPEs"} which is a list of summary statistics of 
\code{object}.
}
\description{
Summary statistics for objects of class \code{"bifeAPEs"}.
}
\seealso{
\code{\link{get_APEs}}
}
