% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.n.R
\docType{data}
\name{time.n}
\alias{time.n}
\title{Computation time with varying N}
\format{A named matrix with 10 rows and 4 columns.}
\usage{
time.n
}
\description{
Results reported in the vignette.
}
\seealso{
\code{\link{bife}}
}
\keyword{datasets}

