% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.t.R
\docType{data}
\name{time.t}
\alias{time.t}
\title{Computation time with varying T}
\format{A named matrix with 10 rows and 4 columns.}
\usage{
time.t
}
\description{
Results reported in the vignette.
}
\seealso{
\code{\link{bife}}
}
\keyword{datasets}

