% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_t.R
\docType{data}
\name{time_t}
\alias{time_t}
\title{Computation time with varying T}
\format{A named matrix with 10 rows and 4 columns.}
\usage{
time_t
}
\description{
Results reported in the vignette.
}
\seealso{
\code{\link{bife}}
}
\keyword{datasets}
