% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.bife.R
\name{vcov.bife}
\alias{vcov.bife}
\title{Extract Covariance Matrix of Structural Parameters}
\usage{
\method{vcov}{bife}(object, corrected = TRUE, ...)
}
\arguments{
\item{object}{an object of class \code{bife}.}

\item{corrected}{an optional logical flag that specifies whether the covariance
matrix of the bias-corrected or uncorrected structural parameters
are displayed. Default is \code{TRUE} (bias-corrected).}

\item{...}{other arguments}
}
\value{
The function \code{vcov.bife} returns a named covariance matrix of the
structural parameters.
}
\description{
\code{vcov.bife} extracts the covariance matrix of the
structural paramters from objects returned by \code{bife}
}
\author{
Amrei Stammann, Daniel Czarnowske, Florian Heiss, Daniel McFadden
}
\seealso{
\code{\link{bife}}
}

