% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{vcov.bifeAPEs}
\alias{vcov.bifeAPEs}
\title{Extract estimates of the covariance matrix}
\usage{
\method{vcov}{bifeAPEs}(object, ...)
}
\arguments{
\item{object}{an object of class \code{"bifeAPEs"}.}

\item{...}{other arguments.}
}
\value{
The function \code{\link{vcov.bifeAPEs}} returns a named matrix of covariance estimates.
}
\description{
\code{\link{vcov.bifeAPEs}} computes an estimate of the covariance matrix of the estimator of the
average partial parameters from objects returned by \code{\link{get_APEs}}.
}
\seealso{
\code{\link{get_APEs}}
}
