% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{coef.bife}
\alias{coef.bife}
\title{Extract estimates of structural parameters or fixed effects}
\usage{
\method{coef}{bife}(object, type = c("sp", "fe"), corrected = NULL, fixed = NULL, ...)
}
\arguments{
\item{object}{an object of class \code{"bife"}.}

\item{type}{the type of parameter estimates that should be returned; structural parameters or 
fixed effects. Default is \code{"sp"} referring to the structural parameters.}

\item{corrected, fixed}{deprecated.}

\item{...}{other arguments.}
}
\value{
The function \code{\link{coef.bife}} returns a named vector of estimates of the requested 
parameters.
}
\description{
\code{\link{coef.bife}} is a generic function which extracts estimates of the structural 
parameters or fixed effects from objects returned by \code{\link{bife}}.
}
\seealso{
\code{\link{bife}}
}
