% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{print.bife}
\alias{print.bife}
\title{Print \code{bife}}
\usage{
\method{print}{bife}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{an object of class \code{"bife"}.}

\item{digits}{unsigned integer indicating the number of decimal places. Default is 
\code{max(3L, getOption("digits") - 3L)}.}

\item{...}{other arguments.}
}
\description{
\code{\link{print.bife}} is a generic function which displays some minimal information from 
objects returned by \code{\link{bife}}.
}
\seealso{
\code{\link{bife}}
}
