% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{print.summary.bife}
\alias{print.summary.bife}
\title{Print \code{summary.bife}}
\usage{
\method{print}{summary.bife}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{an object of class \code{"summary.bife"}.}

\item{digits}{unsigned integer indicating the number of decimal places. Default is 
\code{max(3L, getOption("digits") - 3L)}.}

\item{...}{other arguments.}
}
\description{
\code{\link{print.summary.bife}} is a generic function which displays summary statistics from 
objects returned by \code{\link{summary.bife}}.
}
\seealso{
\code{\link{bife}}
}
