% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{summary.bife}
\alias{summary.bife}
\title{Summarizing models of class \code{bife}}
\usage{
\method{summary}{bife}(object, type = c("sp", "fe"), corrected = NULL, fixed = NULL, ...)
}
\arguments{
\item{object}{an object of class \code{"bife"}.}

\item{type}{the type of parameter estimates the summary statistics are related to: structural 
parameters or fixed effects. Default is \code{"sp"} referring to the structural parameters.}

\item{corrected, fixed}{deprecated.}

\item{...}{other arguments.}
}
\value{
Returns an object of class \code{"summary.bife"} which is a list of summary statistics of
\code{object}.
}
\description{
Summary statistics for objects of class \code{"bife"}.
}
\seealso{
\code{\link{bife}}
}
