% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{vcov.bife}
\alias{vcov.bife}
\title{Extract estimates of the covariance matrix}
\usage{
\method{vcov}{bife}(object, ...)
}
\arguments{
\item{object}{an object of class \code{"bife"}.}

\item{...}{other arguments.}
}
\value{
The function \code{\link{vcov.bife}} returns a named matrix of covariance estimates.
}
\description{
\code{\link{vcov.bife}} computes an estimate of the covariance matrix of the estimator of the
structural parameters from objects returned by \code{\link{bife}}. The estimate is obtained
using the inverse of the negative Hessian after convergence.
}
\seealso{
\code{\link{bife}}
}
